Refactored app initialization for test setup
