+++ server/app.js
@@ -1,0 +1,28 @@
+import dotenv from "dotenv";
+import express from "express";
+import cors from "cors";
+import bodyParser from "body-parser";
+import cookieParser from "cookie-parser";
+import routes from "./routes.js";
+
+/* CONFIGURATIONS */
+dotenv.config();
+const app = express();
+app.use(express.json());
+app.use(bodyParser.json({extended: true}));
+app.use(bodyParser.urlencoded({extended: true}));
+
+// app.use(cors());
+
+app.use((req, res, next) => {
+    res.setHeader("Access-Control-Allow-Origin", "http://localhost:3000");
+    res.setHeader("Access-Control-Allow-Methods", "POST");
+    res.setHeader("Access-Control-Allow-Headers", "Access-Control-Allow-Headers,Access-Control-Allow-Methods,Origin,Accept,Content-Type");
+    res.setHeader("Access-Control-Allow-Credentials","true");
+    next();
+});
+
+app.use(cookieParser());
+app.use(routes);
+
+export default app;
