+++ server/index.js
@@ -1,33 +1,9 @@
-import express from "express";
-import cors from "cors";
-import dotenv from "dotenv";
-import bodyParser from "body-parser";
-import mongoose from "mongoose";
-import cookieParser from "cookie-parser";
+import app from './app.js';
+import dotenv from 'dotenv';
 import makeDB from "./mongoose.js";
-import routes from "./routes.js";
-
-/* CONFIGURATIONS */
-dotenv.config();
-const app = express();
-app.use(express.json());
-app.use(bodyParser.json({extended: true}));
-app.use(bodyParser.urlencoded({extended: true}));
-
-// app.use(cors());
-
-app.use((req, res, next) => {
-    res.setHeader("Access-Control-Allow-Origin", "http://localhost:3000");
-    res.setHeader("Access-Control-Allow-Methods", "POST");
-    res.setHeader("Access-Control-Allow-Headers", "Access-Control-Allow-Headers,Access-Control-Allow-Methods,Origin,Accept,Content-Type");
-    res.setHeader("Access-Control-Allow-Credentials","true");
-    next();
-});
-
-app.use(cookieParser());
-app.use(routes);
 
 /* DB SETUP */
-const PORT = process.env.PORT || 6001;
+dotenv.config();
+const port = process.env.PORT || 6001;
 
-makeDB(process.env.MONGO_URL).then(() => app.listen(PORT, () => console.log(`SERVER PORT: ${PORT}`)))
+makeDB(process.env.MONGO_URL).then(() => app.listen(port, () => console.log(`SERVER PORT: ${port}`)))
